Moved temperature entry class definition to a seperate file
