+++ archey/archey.py
@@ -6,7 +6,6 @@ import re
 import sys
 
 from enum import Enum
-from glob import glob
 from subprocess import CalledProcessError, DEVNULL, check_output
 
 import distro
@@ -121,71 +120,6 @@ class Terminal:
         self.value = '{0} {1}'.format(terminal, colors)
 
 
-class Temperature:
-    def __init__(self):
-        temps = []
-
-        try:
-            # Let's try to retrieve a value from the Broadcom chip on Raspberry
-            temp = float(
-                re.search(
-                    r'\d+\.\d+',
-                    check_output(
-                        ['/opt/vc/bin/vcgencmd', 'measure_temp'],
-                        stderr=DEVNULL, universal_newlines=True
-                    )
-                ).group(0)
-            )
-
-            temps.append(
-                self._convert_to_fahrenheit(temp)
-                if CONFIG.get('temperature')['use_fahrenheit'] else temp
-            )
-
-        except (FileNotFoundError, CalledProcessError):
-            pass
-
-        # Now we just check for values within files present in the path below
-        for thermal_file in glob('/sys/class/thermal/thermal_zone*/temp'):
-            with open(thermal_file) as file:
-                try:
-                    temp = float(file.read().strip()) / 1000
-
-                except OSError:
-                    continue
-
-                if temp != 0.0:
-                    temps.append(
-                        self._convert_to_fahrenheit(temp)
-                        if CONFIG.get('temperature')['use_fahrenheit']
-                        else temp
-                    )
-
-        if temps:
-            self.value = '{0}{1}{2}'.format(
-                str(round(sum(temps) / len(temps), 1)),
-                CONFIG.get('temperature')['char_before_unit'],
-                'F' if CONFIG.get('temperature')['use_fahrenheit'] else 'C'
-            )
-
-            if len(temps) > 1:
-                self.value += ' (Max. {0}{1}{2})'.format(
-                    str(round(max(temps), 1)),
-                    CONFIG.get('temperature')['char_before_unit'],
-                    'F' if CONFIG.get('temperature')['use_fahrenheit'] else 'C'
-                )
-
-        else:
-            self.value = CONFIG.get('default_strings')['not_detected']
-
-    @staticmethod
-    def _convert_to_fahrenheit(temp):
-        """
-        Simple Celsius to Fahrenheit conversion method
-        """
-        return temp * (9 / 5) + 32
-
-
 # ----------- Classes Index ----------- #
 
 class Classes(Enum):
@@ -210,7 +144,14 @@ class Classes(Enum):
             'not_detected': CONFIG.get('default_strings')['not_detected']
         }
     }
-    Temperature = {'class': Temperature}
+    Temperature = {
+        'class': entries.Temperature,
+        'kwargs': {
+            'use_fahrenheit': CONFIG.get('temperature')['use_fahrenheit'],
+            'char_before_unit': CONFIG.get('temperature')['char_before_unit'],
+            'not_detected': CONFIG.get('default_strings')['not_detected']
+        }
+    }
     CPU = {'class': entries.CPU}
     GPU = {
          'class': entries.GPU,
