+++ archey/entries/__init__.py
@@ -4,6 +4,7 @@ from .model import Model
 from .kernel import Kernel
 from .uptime import Uptime
 from .packages import Packages
+from .temperature import Temperature
 from .cpu import CPU
 from .gpu import GPU
 from .ram import RAM
